tr2(1)

+++ tasks/mpi/vladimirova_j_gather/src/ops_mpi.cpp
@@ -22,14 +22,14 @@ std::vector<int> vladimirova_j_gather_mpi::getRandomVector(int sz) {
 
 std::vector<int> vladimirova_j_gather_mpi::noDeadEnds(std::vector<int> way) {
   int i = 0;
-  int j = 1;
+  size_t j = 1;
 
   while (j <= way.size()) {
     if ((way[i] * way[i] == 1) && (way[i] == way[j])) {
       do {
         i -= 1;
         j += 1;
-        if (!((i >= 0) && (j < way.size()))) break;
+        if (!(((size_t)i >= 0) && (j < way.size()))) break;
         if (((way[i] * way[i] == 1) && (way[i] == (-1) * way[j])) ||
             (way[i] * way[j] == 4)) {  // if rl lr or uu dd   1-1 -11 or 22 -2-2
           way[i] = 0;
@@ -57,14 +57,14 @@ std::vector<int> vladimirova_j_gather_mpi::noStrangeSteps(std::vector<int> way) 
     std::cout << value << " ";
   }
 
-  for (int i = 1; i < way.size(); i++)
+  for (size_t i = 1; i < way.size(); i++)
     if (way[i] == -way[i - 1]) {
       way[i] = 0;
       way[i - 1] = 0;
     }
   way.erase(std::remove(way.begin(), way.end(), 0), way.end());
 
-  for (int i = 3; i < way.size(); i++) {
+  for (size_t i = 3; i < way.size(); i++) {
     if (((way[i] == -1) || (way[i] == 1)) && (way[i] == way[i - 1]) && (way[i] == way[i - 2]) &&
         (way[i] == way[i - 3])) {
       way[i] = 0;
@@ -94,14 +94,14 @@ std::vector<int> vladimirova_j_gather_mpi::convertToBinaryTreeOrder(const std::v
     int r = stack.back();
     stack.pop_back();
 
-    if (r < arr.size()) {
+    if ((size_t)r < arr.size()) {
       result.push_back(arr[r]);
 
       int child1 = 2 * r + 2;
       int child0 = 2 * r + 1;
 
-      if (child1 < arr.size()) stack.push_back(child1);
-      if (child0 < arr.size()) stack.push_back(child0);
+      if ((size_t)child1 < arr.size()) stack.push_back(child1);
+      if ((size_t)child0 < arr.size()) stack.push_back(child0);
     }
   }
 
@@ -112,7 +112,7 @@ bool vladimirova_j_gather_mpi::TestMPITaskSequential::pre_processing() {
   internal_order_test();
   input_ = std::vector<int>(taskData->inputs_count[0]);
   auto* tmp_ptr = reinterpret_cast<int*>(taskData->inputs[0]);
-  for (unsigned i = 0; i < taskData->inputs_count[0]; i++) {
+  for (size_t i = 0; i < taskData->inputs_count[0]; i++) {
     input_[i] = tmp_ptr[i];
   }
   return true;
@@ -146,7 +146,7 @@ bool vladimirova_j_gather_mpi::TestMPITaskParallel::pre_processing() {
     // Init vectors
     input_ = std::vector<int>(taskData->inputs_count[0]);
     auto* tmp_ptr = reinterpret_cast<int*>(taskData->inputs[0]);
-    for (unsigned i = 0; i < taskData->inputs_count[0]; i++) {
+    for (size_t i = 0; i < taskData->inputs_count[0]; i++) {
       input_[i] = tmp_ptr[i];
     }
   }
@@ -226,12 +226,12 @@ bool vladimirova_j_gather_mpi::TestMPITaskParallel::run() {
 
   if (r == 0) {
     size = input_.size() / world.size();
-    for (int i = 1; i < world.size(); i++) {
+    for (size_t i = 1; i < world.size(); i++) {
       world.send(i, 0, size);
     }
 
     std::vector<int> pr(world.size());
-    for (int i = 0; i < pr.size(); i++) {
+    for (size_t i = 0; i < pr.size(); i++) {
       pr[i] = i;
     }
 
@@ -245,7 +245,7 @@ bool vladimirova_j_gather_mpi::TestMPITaskParallel::run() {
     std::for_each(pr.begin(), pr.end(), [](int number) { std::cout << number << " "; });
     std::cout << std::endl;
 
-    for (int i = 1; i < world.size(); i++) {
+    for (size_t i = 1; i < world.size(); i++) {
       world.send(pr[i], 0, input_.data() + size * i, size);
     }
